Separate out entity models in SOFT and DLite
Create a "minimum set" of fields for the versioned SOFT models and then
create the DLite implementations of these separately.
