+++ tests/models/test_init.py
@@ -15,8 +15,7 @@ def test_soft_entity(static_dir: Path) -> None:
     import json
 
     from dlite_entities_service.models import soft_entity
-    from dlite_entities_service.models.soft5 import SOFT5Entity
-    from dlite_entities_service.models.soft7 import SOFT7Entity
+    from dlite_entities_service.models.dlite import DLiteSOFT5Entity, DLiteSOFT7Entity
 
     # Test that the function returns the correct version of the entity
     soft5_model_file = static_dir / "valid_entities" / "Cat.json"
@@ -25,8 +24,8 @@ def test_soft_entity(static_dir: Path) -> None:
     soft5_model = json.loads(soft5_model_file.read_text())
     soft7_model = json.loads(soft7_model_file.read_text())
 
-    assert soft_entity(**soft5_model) == SOFT5Entity(**soft5_model)
-    assert soft_entity(**soft7_model) == SOFT7Entity(**soft7_model)
+    assert soft_entity(**soft5_model) == DLiteSOFT5Entity(**soft5_model)
+    assert soft_entity(**soft7_model) == DLiteSOFT7Entity(**soft7_model)
 
 
 def test_soft_entity_error(static_dir: Path) -> None:
@@ -36,6 +35,7 @@ def test_soft_entity_error(static_dir: Path) -> None:
     from pydantic import ValidationError
 
     from dlite_entities_service.models import soft_entity
+    from dlite_entities_service.models.dlite import DLiteSOFT5Entity, DLiteSOFT7Entity
     from dlite_entities_service.models.soft5 import SOFT5Entity
     from dlite_entities_service.models.soft7 import SOFT7Entity
 
@@ -49,6 +49,12 @@ def test_soft_entity_error(static_dir: Path) -> None:
     errors = soft_entity(return_errors=True, **invalid_model)
 
     expected_errors = []
+    # The order here is important, as it represents the order in which the models
+    # are tried in the soft_entity function.
+    # The order is defined by the Union arguments in the Entity type:
+    # Entity = SOFT7Entity | SOFT5Entity | DLiteEntity
+    # And again the order of the Union arguments in the DLiteEntity type:
+    # DLiteEntity = DLiteSOFT7Entity | DLiteSOFT5Entity
     try:
         SOFT7Entity(**invalid_model)
     except ValidationError as exc:
@@ -56,6 +62,16 @@ def test_soft_entity_error(static_dir: Path) -> None:
 
     try:
         SOFT5Entity(**invalid_model)
+    except ValidationError as exc:
+        expected_errors.append(exc)
+
+    try:
+        DLiteSOFT7Entity(**invalid_model)
+    except ValidationError as exc:
+        expected_errors.append(exc)
+
+    try:
+        DLiteSOFT5Entity(**invalid_model)
     except ValidationError as exc:
         expected_errors.append(exc)
